requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
assemblyai==0.37.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
Django==4.2.20
exceptiongroup==1.2.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.68.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
pytube==15.0.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==15.0.1
yt-dlp==2025.3.26
